Update example
We don't need to test the `jax.grad` of this function directly as that does not make much sense in this context.  It is not a scalar valued function with a scalar valued output.

Note: the `jax.jacfwd` of the deflection angles *is* an important value, but that is a different example all together.