Example for taking the hessian of deflection angles
This example shows how to compute the area distortion matrix of a lems mapping.  This can be used to find the kappa, sheer, and magnifications from a deflection angle function.

At the moment it only returns a JAX array, these results still need to be wrapped into a vector2D class I think.